// src/app/generated_services/model/carlonGracieBackendProgressionApplicationDTOsShowBeltDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CarlonGracieBackendProgressionApplicationDTOsShowBeltDTO { 
    id?: string;
    color: string;
    orderIndex?: number;
    isForKids?: boolean;
}