// src/app/generated_services/model/carlonGracieBackendStudentsApplicationDTOsShowStudentDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CarlonGracieBackendStudentsApplicationDTOsShowStudentDTO { 
    id?: string;
    preferredUsername?: string | null;
    photoUrl?: string | null;
    createdAt?: string;
    lastLoginAt?: string | null;
    totalFrequencies?: number;
    totalGraduations?: number;
    currentBelt?: string | null;
    userName: string;
    email: string;
    phoneNumber?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    birthDay?: string | null;
    isActive?: boolean;
}